move streaming function into its own module
